requirements.txt
pyyaml~=5.3
Flask~=1.1.2
pytest~=6.1.2
pandas~=1.1.5
urllib3~=1.25.11
geopy~=2.2.0
python-dotenv~=0.17.1
scikit-learn~=0.24.2
numpy~=1.21.2
keras~=2.6.0
tensorflow~=2.6.0